backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg[binary]>=3.2.0
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.74
langchain>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.27.0
airbyte-agent-linear>=0.19.121
